Split hardcoded GraphQL types into separate files
